+++ model/tests/CPU.memory.test.cpp
@@ -3,7 +3,7 @@
 
 #include <catch.hpp>
 
-#include "mock/Aux.hpp"
+#include "mock/Helper.hpp"
 #include "model/CPU.hpp"
 
 namespace
